backend/pyproject.toml
[tool.poetry]
name = "mssite"
version = "0.1.0"
description = "backend part for MIS_EXPRESS site"
authors = ["zormnus", "kuzilya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
pillow = "^10.2.0"
python-slugify = "^8.0.4"
drf-spectacular = "^0.27.1"
drf-spectacular-sidecar = "^2024.2.1"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.31.0"
faker = "^23.2.1"
django-debug-toolbar = "^4.3.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
factory-boy = "^3.3.0"
django-import-export = "^4.0.9"
openpyxl = "^3.1.4"
django-filter = "^24.2"
slugify = "^0.0.1"
django-parler = "^2.3"
django-parler-rest = "^2.2"
deepl = "^1.18.0"
webcolors = "^24.6.0"
googletrans = "3.1.0a0"
celery = "^5.4.0"
redis = "^5.0.7"
billiard = "^4.2.0"
kombu = "^5.3.7"
django-mptt = "^0.16.0"
twilio = "^9.2.3"
beautifulsoup4 = "^4.12.3"
vonage = "^3.16.1"
messagebird = "^2.2.0"
smsaero-api = "^3.0.0"
django-sql-debug = "^0.0.3"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
